\name{eff_dat}
\docType{data}
\alias{eff_dat}
\title{Patient efficacy data}
\description{
  This data set provides the efficacy data from 22 patients. 
}
\usage{eff_dat}
\format{A matrix containing 22 observations.}
\source{Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.}
\references{
  Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.
}
\keyword{datasets}
