\name{patdata}
\docType{data}
\alias{patdata}
\title{Patient toxicity data}
\description{
  This data set provides the toxicity data from multiple treatment cycles. The toxicity data is defined by nTTP toxicity score. 
}
\usage{patdata}
\format{A data frame containing 9 observations.}
\source{Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.}
\references{
  Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.
}
\keyword{datasets}
