\name{probaT}
\docType{data}
\alias{probaT}
\title{The probability matrix of toxicity events}
\description{
  This data set provides the toxicity matrix of 3 types of toxicity events over 6 doses and 5 treatment cycles. 
}
\usage{probaT}
\format{A list containing 3 matrices for the probability for 3 type of toxicity events, each matrix with 6 rows for 6 doses and 5 columns for 5 treatment cycles.}
\source{Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.}
\references{
  Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.
}
\keyword{datasets}
