\name{tox_dat}
\docType{data}
\alias{tox_dat}
\title{Patient toxicity data}
\description{
  This data set provides the toxicity data from 9 patients with multiple treatment cycles, matching with the efficacy data. 
}
\usage{tox_dat}
\format{A data frame containing 183 observations.}
\source{Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.}
\references{
  Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.
}
\keyword{datasets}
