\name{tox.matrix}
\docType{data}
\alias{tox.matrix}
\title{Probability matrix for toxicity}
\description{
  This data set is a four-dimension array (Dose, Cycle, Toxicity Type, Toxicity Grade) specifying the probabilities of the occurrences of certain grades for certain types of toxicities.
}
\usage{tox.matrix}
\format{A 4-dimension array.}
\source{Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.}
\references{
  Yin, Jun, et al. "A Bayesian dose finding design incorporating toxicity data from multiple treatment cycles." Statistics in Medicine 36.1 (2017): 67-80.
}
\keyword{datasets}
